requirements.txt

streamlit
pandas
bcrypt
psycopg2-binary
